Código comentado, colisão arrumada, lógica simplificada

+++ Projeto 1904/personagens.h
@@ -4,11 +4,11 @@
 #include <stdbool.h>
 
 typedef struct {
-	float posicaoXp, posicaoYp, largurap, alturap;
+	int posicaoXp, posicaoYp, largurap, alturap;
 }Jogador;
 
 typedef struct {
-	float posicaoX, posicaoY, largura, altura;
+	int posicaoX, posicaoY, largura, altura;
 }Inimigo;
 
 // Funções jogador
